src/mcp_common/pyproject.toml
[project]
name = "mcp_common"
version = "0.1.0"
description = "Shared MCP utilities"
authors = [{ name = "Jaroslav" }]
dependencies = [
    "fastapi",
    "boto3",
    "python-dotenv"
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
